% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BAV.R
\name{BAV}
\alias{BAV}
\title{BAV}
\usage{
BAV(deploy = TRUE, model_units = 1)
}
\arguments{
\item{deploy}{boolean, TRUE if the model should be deployed in a live trading environment, FALSE if it's just being used for testing/research, Default: TRUE}

\item{model_units}{numeric, how many units to allocate to this model}
}
\value{
An xts of units to be combined with an unconditional return series to form portfolio (ie condition) returns.
}
\description{
Below AVerage.
}
\details{
Enter at close and hold for 24 hours if today's return is below its SMA (lookback length being optimized monthly using last year's data).
}
