% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CAN.R
\name{CAN}
\alias{CAN}
\title{CAN}
\usage{
CAN(deploy = TRUE, model_units = 1.5)
}
\arguments{
\item{deploy}{boolean, TRUE if the model should be deployed in a live trading environment, FALSE if it's just being used for testing/research, Default: TRUE}

\item{model_units, }{how many units to allocate to this model}
}
\value{
An xts of units to be combined with an unconditional return series to form portfolio (ie condition) returns.
}
\description{
CANary.
}
\details{
This is the KDA algorithm (reference below) but only ever goes long SPY.
}
\references{
\url{https://quantstrattrader.com/2019/02/27/kda-robustness-results/}
}
