% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KDA.R
\name{KDA}
\alias{KDA}
\title{KDA}
\usage{
KDA(deploy = TRUE, model_units = 1.5, units_fun = "units_kda")
}
\arguments{
\item{deploy}{boolean, TRUE if the model should be deployed in a live trading environment, FALSE if it's just being used for testing/research, Default: TRUE}

\item{model_units}{numeric, the number of units to allocate to this model, Default: 1.5}

\item{units_fun}{character, the KDA variant's function name, Default: 'units_kda'}
}
\value{
An xts of units to be combined with an unconditional return series to form portfolio (ie condition) returns.
}
\description{
Kipnis Defensive Adaptive Asset Allocation.
}
\details{
See References.
}
\references{
\url{https://quantstrattrader.com/2019/02/27/kda-robustness-results/}
}
