% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VIM.R
\name{VIM}
\alias{VIM}
\title{VIM}
\usage{
VIM(deploy = TRUE, model_units = 1)
}
\arguments{
\item{deploy}{boolean, TRUE if the model should be deployed in a live trading environment, FALSE if it's just being used for testing/research, Default: TRUE}

\item{model_units, }{how many units to allocate to this model}
}
\value{
An xts of units to be combined with an unconditional return series to form portfolio (ie condition) returns.
}
\description{
VIx Momentum
}
\details{
The same momentum sorter from CAN selects 1 best and 1 worst Dow component for the month.  Long entry determined by EMA-based VIX filter.
}
